Extract SSHFileExplorerPathResolver to its own file
Moves the three pure path-resolution helpers
(remoteHomePath, effectiveRootPath, isMacLocalPath) out of
ContentView.swift and into their own file so they sit at a
package boundary and can be exercised without instantiating an
NSApplication or a Workspace graph.

Also marks all three helpers `nonisolated` so a future
module-level default actor isolation setting cannot silently
re-bind them to the main actor, and drops the now-unused
`remotePath` binding in syncFileExplorerDirectory whose
flatMap consumer was replaced by the resolver call.
