+++ Sources/ContentView.swift
@@ -7,78 +7,6 @@ import ObjectiveC
 import UniformTypeIdentifiers
 import WebKit
 
-/// Pure helpers used by `ContentView.syncFileExplorerDirectory` to compute
-/// the remote home directory and effective root path for the file explorer
-/// when a workspace is bound to an SSH remote.
-///
-/// Lives at top level (rather than nested in ContentView) so the cmuxTests
-/// target can reach it via `@testable import cmux`.
-enum SSHFileExplorerPathResolver {
-    /// Derives the conventional remote home directory from the SSH
-    /// `<user>@<host>` destination string.
-    ///
-    /// Linux convention: `$HOME` is `/home/<user>`, except `root` which uses
-    /// `/root`. Returns "" when the destination is nil/empty or the user
-    /// portion cannot be confidently extracted (e.g. missing `@`, whitespace
-    /// only). Whitespace around the destination and around the user token is
-    /// trimmed; ports/IPv6 brackets after the host are ignored.
-    ///
-    /// Examples:
-    ///     "imgyu@100.79.206.23"  -> "/home/imgyu"
-    ///     "ubuntu@host:2222"      -> "/home/ubuntu"
-    ///     "root@host"             -> "/root"
-    ///     "user@[::1]:22"         -> "/home/user"
-    ///     ""                      -> ""
-    ///     "noatsign"              -> ""   (no user portion is reliable)
-    ///     "@host"                 -> ""
-    ///     "user@"                 -> ""   (no host = malformed)
-    static func remoteHomePath(from destination: String?) -> String {
-        guard let raw = destination?.trimmingCharacters(in: .whitespacesAndNewlines),
-              !raw.isEmpty else {
-            return ""
-        }
-        let atParts = raw.split(separator: "@", maxSplits: 1, omittingEmptySubsequences: false)
-        guard atParts.count == 2 else { return "" }
-        let user = atParts[0].trimmingCharacters(in: .whitespacesAndNewlines)
-        let hostPart = atParts[1].trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !user.isEmpty, !hostPart.isEmpty else { return "" }
-        if user == "root" { return "/root" }
-        return "/home/\(user)"
-    }
-
-    /// Returns the file-tree root path to use for an SSH-bound workspace.
-    ///
-    /// `workspaceDirectory` is the workspace's tracked cwd. For SSH workspaces
-    /// created on macOS this is typically the Mac caller's cwd
-    /// (e.g. `/Users/<x>/Downloads`) and is unreachable on the remote, so we
-    /// fall back to the derived remote home when:
-    ///     - `remoteHomePath` is non-empty (we have a credible anchor), AND
-    ///     - `workspaceDirectory` looks like a Mac-local path (`/Users/...`,
-    ///       `/Volumes/...`).
-    /// In every other case we return the workspace directory unchanged so we
-    /// don't accidentally redirect a user who is intentionally rooted at an
-    /// already-remote-style path (`/home/...`, `/root`, `/etc`, `/`, etc.).
-    static func effectiveRootPath(
-        workspaceDirectory: String,
-        remoteHomePath: String
-    ) -> String {
-        guard !remoteHomePath.isEmpty else { return workspaceDirectory }
-        if isMacLocalPath(workspaceDirectory) {
-            return remoteHomePath
-        }
-        return workspaceDirectory
-    }
-
-    /// Returns true if `path` is on a macOS-only mount that cannot exist on a
-    /// Linux remote: `/Users/...` (home) or `/Volumes/...` (mounted disks).
-    /// We deliberately do NOT match `/private/...`, `/tmp/...`, `/etc/...` and
-    /// other roots that exist on both platforms — those may legitimately be
-    /// the user's intended remote target.
-    static func isMacLocalPath(_ path: String) -> Bool {
-        path.hasPrefix("/Users/") || path.hasPrefix("/Volumes/")
-    }
-}
-
 /// Transparent NSView installed on the window's theme frame (above the NSHostingView) to
 /// handle file/URL drags from Finder. Nested NSHostingController layers (created by bonsplit's
 /// SinglePaneWrapper) prevent AppKit's NSDraggingDestination routing from reaching deeply
@@ -3106,7 +3034,6 @@ struct ContentView: View {
 
         if tab.isRemoteWorkspace {
             let config = tab.remoteConfiguration
-            let remotePath = tab.remoteDaemonStatus.remotePath
             let isReady = tab.remoteDaemonStatus.state == .ready
             // See SSHFileExplorerPathResolver for the rationale + test cases.
             let homePath = SSHFileExplorerPathResolver.remoteHomePath(from: config?.destination)
